+++ stroom-cache/stroom-cache-impl/src/main/java/stroom/cache/impl/CacheManagerService.java
@@ -33,4 +33,6 @@ public interface CacheManagerService {
     Long clear(FindCacheInfoCriteria criteria);
 
     void evictExpiredElements();
+
+    Long evictExpiredElements(FindCacheInfoCriteria criteria);
 }